+++ lib/td_dq/rules/rule_result.ex
@@ -13,9 +13,12 @@ defmodule TdDq.Rules.RuleResult do
   alias TdDq.Implementations.Implementation
   alias TdDq.Remediations.Remediation
   alias TdDq.Rules.Rule
+  alias TdDq.Rules.RuleResult
 
   @scale 2
   @valid_result_types Implementation.valid_result_types()
+  @segments_validations [:result, :date, :result_type, :parent_id]
+  @rule_resuls_validations [:implementation, :date, :result, :result_type, :rule_id]
 
   @typedoc "The result of a quality rule execution"
   @type t :: %__MODULE__{}
@@ -32,6 +35,7 @@ defmodule TdDq.Rules.RuleResult do
 
     belongs_to(:implementation, Implementation)
     belongs_to(:rule, Rule)
+    belongs_to(:parent, RuleResult)
 
     has_many(:execution, Execution, foreign_key: :result_id)
     has_one(:remediation, Remediation)
@@ -52,18 +56,25 @@ defmodule TdDq.Rules.RuleResult do
       :params,
       :row_number,
       :result_type,
-      :details
+      :details,
+      :parent_id
     ])
-    |> put_assoc(:implementation, implementation)
+    |> add_assoc(implementation)
     |> put_date()
     |> maybe_put_result()
     |> validate_inclusion(:result_type, @valid_result_types)
     |> update_change(:result, &Decimal.round(&1, @scale, :floor))
-    |> validate_required([:implementation, :date, :result, :result_type, :rule_id])
+    |> add_validations()
     |> validate_number(:records, greater_than_or_equal_to: 0)
     |> validate_number(:errors, greater_than_or_equal_to: 0)
     |> foreign_key_constraint(:rule_id)
   end
+
+  defp add_assoc(%{data: %{parent_id: parent_id}} = changeset, _impl) when not is_nil(parent_id),
+    do: changeset
+
+  defp add_assoc(changeset, implementation),
+    do: put_assoc(changeset, :implementation, implementation)
 
   defp put_date(%{params: %{"date" => value}} = changeset) do
     # Standard datetime formats will be cast correctly by Ecto, we only need to
@@ -78,17 +89,24 @@ defmodule TdDq.Rules.RuleResult do
 
   defp maybe_put_result(%{changes: %{result: _result}} = changeset), do: changeset
 
-  defp maybe_put_result(%{} = changeset) do
+  defp maybe_put_result(changeset) do
     with records when is_integer(records) <- get_change(changeset, :records),
          errors when is_integer(errors) <- get_change(changeset, :errors),
-         result_type when result_type != nil <-
-           changeset.data.result_type || get_change(changeset, :result_type) do
+         result_type when result_type != nil <- get_result_type(changeset) do
       result = calculate_quality(records, errors, result_type)
       put_change(changeset, :result, result)
     else
       _result -> changeset
     end
   end
+
+  defp get_result_type(%{data: %{result_type: result_type}}) when not is_nil(result_type),
+    do: result_type
+
+  defp get_result_type(%{changes: %{result_type: result_type}}) when not is_nil(result_type),
+    do: result_type
+
+  defp get_result_type(changeset), do: get_change(changeset, :result_type)
 
   defp calculate_quality(0, _errors, _result_type), do: 0
 
@@ -109,5 +127,17 @@ defmodule TdDq.Rules.RuleResult do
     |> Decimal.abs()
     |> Decimal.mult(100)
     |> Decimal.div(records)
+  end
+
+  defp add_validations(%{data: %{parent_id: parent_id}} = changeset) when not is_nil(parent_id) do
+    validate_required(changeset, @segments_validations)
+  end
+
+  defp add_validations(%{changes: %{parent_id: _}} = changeset) do
+    validate_required(changeset, @segments_validations)
+  end
+
+  defp add_validations(changeset) do
+    validate_required(changeset, @rule_resuls_validations)
   end
 end