+++ lib/td_dq_web/controllers/implementation_controller.ex
@@ -118,18 +118,26 @@ defmodule TdDqWeb.ImplementationController do
 
     implementation =
       id
+      ## TODO: take this functionality con resutl context to add has_segments
       |> Implementations.get_implementation!(
         enrich: [:source, :links],
-        preload: [:rule, [results: :remediation]]
+        preload: [
+          :rule,
+          [results: :remediation],
+          [data_structures: [data_structure: [:system, :current_version]]]
+        ]
       )
       |> add_last_rule_result()
+      |> with_has_segments()
       |> add_quality_event()
       |> Implementations.enrich_implementation_structures()
       |> filter_links_by_permission(claims)
+      |> filter_data_structures_by_permission(claims)
 
     actions =
       %{}
       |> link_concept_actions(claims, implementation)
+      |> link_structure_actions(claims, implementation)
 
     with {:can, true} <- {:can, can?(claims, show(implementation))} do
       render(conn, "show.json", implementation: implementation, actions: actions)
@@ -158,6 +166,30 @@ defmodule TdDqWeb.ImplementationController do
       actions
     end
   end
+
+  defp link_structure_actions(actions, claims, implementation) do
+    if can?(claims, link_structure(implementation)) do
+      Map.put(actions, :link_structure, %{method: "POST"})
+    else
+      actions
+    end
+  end
+
+  defp filter_data_structures_by_permission(implementation, %{role: "admin"}), do: implementation
+
+  defp filter_data_structures_by_permission(
+         %{data_structures: [_ | _] = data_structures} = implementation,
+         claims
+       ) do
+    data_structures =
+      Enum.filter(data_structures, fn %{data_structure: data_structure} ->
+        can?(claims, show(data_structure))
+      end)
+
+    Map.put(implementation, :data_structures, data_structures)
+  end
+
+  defp filter_data_structures_by_permission(implementation, _claims), do: implementation
 
   swagger_path :update do
     description("Updates Quality Rule")
@@ -287,4 +319,23 @@ defmodule TdDqWeb.ImplementationController do
         params
     end
   end
+
+  ## TODO: refactor this function with SQL sentence
+  defp with_has_segments(%Implementation{results: results} = impl)
+       when length(results) >= 1 do
+    parent_ids =
+      results
+      |> Enum.map(fn %{id: id} -> id end)
+      |> RuleResults.has_segments()
+
+    results =
+      results
+      |> Enum.map(fn %{id: parent_id} = result ->
+        Map.put(result, :has_segments, Enum.member?(parent_ids, parent_id))
+      end)
+
+    Map.put(impl, :results, results)
+  end
+
+  defp with_has_segments(implementation), do: implementation
 end