+++ lib/td_dq_web/views/implementation_view.ex
@@ -6,6 +6,10 @@ defmodule TdDqWeb.ImplementationView do
   alias TdDqWeb.Implementation.DatasetView
   alias TdDqWeb.Implementation.PopulationsView
   alias TdDqWeb.Implementation.RawContentView
+  alias TdDqWeb.Implementation.SegmentsView
+  alias TdDqWeb.Implementation.StructureView
+  alias TdDqWeb.ImplementationStructureView
+
   alias TdDqWeb.RuleResultView
 
   def render("index.json", %{actions: %{} = actions} = assigns) when map_size(actions) > 0 do
@@ -32,6 +36,8 @@ defmodule TdDqWeb.ImplementationView do
   def render("implementation.json", %{
         implementation: %{implementation_type: "raw"} = implementation
       }) do
+    data_structures = Map.get(implementation, :data_structures)
+
     implementation
     |> Map.take([
       :current_business_concept_version,
@@ -62,9 +68,12 @@ defmodule TdDqWeb.ImplementationView do
     |> add_last_rule_results(implementation)
     |> add_quality_event_info(implementation)
     |> add_rule_results(implementation)
+    |> maybe_render_data_structures(data_structures)
   end
 
   def render("implementation.json", %{implementation: implementation}) do
+    data_structures = Map.get(implementation, :data_structures)
+
     implementation
     |> Map.take([
       :current_business_concept_version,
@@ -92,12 +101,14 @@ defmodule TdDqWeb.ImplementationView do
       :validations,
       render_many(implementation.validations, ConditionView, "condition_row.json")
     )
+    |> add_segments(implementation)
     |> add_first_population(implementation)
     |> add_populations(implementation)
     |> add_rule(implementation)
     |> add_quality_event_info(implementation)
     |> add_last_rule_results(implementation)
     |> add_rule_results(implementation)
+    |> maybe_render_data_structures(data_structures)
   end
 
   defp add_first_population(mapping, %{populations: [%{population: population} | _]})
@@ -120,6 +131,16 @@ defmodule TdDqWeb.ImplementationView do
   end
 
   defp add_populations(mapping, _implementation), do: mapping
+
+  defp add_segments(mapping, %{segments: segments}) do
+    mapping
+    |> Map.put(
+      :segments,
+      render_many(segments, SegmentsView, "segments.json")
+    )
+  end
+
+  defp add_segments(mapping, _implementation), do: mapping
 
   defp add_rule(mapping, %{rule: rule}) when map_size(rule) > 0 do
     rule =
@@ -176,18 +197,27 @@ defmodule TdDqWeb.ImplementationView do
     end
   end
 
-  defp add_rule_results(implementation_mapping, implementation) do
+  defp add_rule_results(implementation_mapping, %{results: [_ | _] = results}) do
     rule_results =
-      implementation
-      |> Map.get(:results, [])
+      results
       |> Enum.sort_by(& &1.date, {:desc, DateTime})
       |> Enum.map(&render_one(&1, RuleResultView, "rule_result.json"))
 
-    case rule_results do
-      [] -> implementation_mapping
-      _ -> Map.put(implementation_mapping, :results, rule_results)
-    end
+    Map.put(implementation_mapping, :results, rule_results)
   end
+
+  defp add_rule_results(implementation_mapping, _), do: implementation_mapping
+
+  defp maybe_render_data_structures(implementation_mapping, data_structures)
+       when is_list(data_structures) do
+    Map.put(
+      implementation_mapping,
+      :data_structures,
+      render_many(data_structures, ImplementationStructureView, "implementation_structure.json")
+    )
+  end
+
+  defp maybe_render_data_structures(implementation_mapping, _), do: implementation_mapping
 end
 
 defmodule TdDqWeb.Implementation.RawContentView do
@@ -321,6 +351,18 @@ defmodule TdDqWeb.Implementation.ConditionView do
   end
 
   defp with_value_modifier(condition, _row), do: condition
+end
+
+defmodule TdDqWeb.Implementation.SegmentsView do
+  use TdDqWeb, :view
+
+  alias TdDqWeb.Implementation.StructureView
+
+  def render("segments.json", %{segments: row}) do
+    %{
+      structure: render_one(row.structure, StructureView, "structure.json")
+    }
+  end
 end
 
 defmodule TdDqWeb.Implementation.OperatorView do